src/api/requirements.txt
amqp==2.3.2
astroid==1.6.3
billiard==3.5.0.4
botocore==1.11.4
celery==4.2.1
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
django==2.0.9
django-admin-sortable2==0.6.19
django-allauth==0.35.0
django-appconf==1.0.2
django-avatar==4.1.0
django-cache-machine-2.0==1.0.0
django-celery-email==2.0.1
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-logging-json==1.15
django-model-utils==3.1.2
django-rest-auth==0.9.3
djangorestframework==3.8.2
docutils==0.14
elasticsearch==6.2.0
factory-boy==2.11.1
Faker==0.8.17
gunicorn==19.7.1
idna==2.6
isort==4.3.4
jmespath==0.9.3
kombu==4.2.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.7
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycurl==7.43.0.2
python-dateutil==2.7.3
python-memcached==1.59
python3-openid==3.1.0
pytz==2018.4
raven==6.9.0
requests==2.20.0
requests-oauthlib==0.8.0
rest-framework-cache==0.1
s3transfer==0.1.13
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1
wrapt==1.10.11
django-storages==1.7.1
boto3==1.9.12
django-admin-list-filter-dropdown==1.0.2
